major refactor part 1 - move all score tracking to bot_data and remove individual chat-based score updates.
This completely removes the need to add chat-based jobs to the job queue so no need to call /start anymore
tracks what teams are 'subscribed to' in chats, and maintains a single set of scores for each of these even if
a team is tracked in multiple chats.
Score updates and milestome reporting are now chained directly to the job that updates the data files so this
is now more timely and more efficient (since there's no need to periodically poll the last update timestamp).